web/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
